Add a workflow using an external action
This should have the same effect as before, but by invoking code in a separate repo. If this works, it will enable us to manage our github actions in a more sustainable way.

I'm not sure about the configuration here - i.e. whether I need to specify the GITHUB_TOKEN here,
or in the action repo (or not at all). I'll figure it out.